requirements.txt
python-telegram-bot==13.7
Flask==2.0.1
python-dotenv==0.20.0
Werkzeug==2.0.3
 Jinja2==3.0.3
